test(safety): guard destructive ops to local test databases

+++ server/test/job.model.test.js
@@ -2,6 +2,7 @@ const test = require('node:test');
 const assert = require('node:assert/strict');
 const mongoose = require('mongoose');
 const { MongoMemoryServer } = require('mongodb-memory-server');
+const { assertSafeTestConnection } = require('./helpers/safe-test-db');
 
 const { Job } = require('../src/models/job');
 
@@ -10,14 +11,18 @@ let mongoServer;
 test.before(async () => {
   mongoServer = await MongoMemoryServer.create();
   await mongoose.connect(mongoServer.getUri(), { dbName: 'xdl_model_test' });
+  assertSafeTestConnection(mongoose.connection);
 });
 
 test.after(async () => {
   await mongoose.disconnect();
-  await mongoServer.stop();
+  if (mongoServer) {
+    await mongoServer.stop();
+  }
 });
 
 test.beforeEach(async () => {
+  assertSafeTestConnection(mongoose.connection);
   const jobsCollection = mongoose.connection.collections.jobs;
   if (jobsCollection) {
     await jobsCollection.deleteMany({});
